murano-apps/package-generator/requirements.txt
GitPython>=1.0.1 # BSD License (3 clause)
pbr<2.0,>=1.6
testresources
python-keystoneclient==1.8.0
python-glanceclient==1.1.0

oslo.config>=2.3.0
oslo.i18n>=1.5.0
oslo.serialization>=1.6.0
oslo.utils>=2.0.0
git+https://github.com/openstack/murano
python-muranoclient
python-heatclient
webob
mock
requests==2.5.3
xmldict==0.4.1
xmltodict==0.9.2
oslo.log
